simple-vision-rag/embedding_service/pyproject.toml
[project]
name = "vision-rag"
version = "0.1.0"
description = "ModernVBERT Multi-Vector Embedding Service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colpali-engine",
    "fastapi>=0.118.0",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.4",
    "qdrant-client>=1.15.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.37.0",
]

[tool.uv.sources]
colpali-engine = { git = "https://github.com/illuin-tech/colpali.git", branch = "vbert" }

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
]


[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
